[055] contactsEventListener + refactor

+++ lib/services/conversation_service.dart
@@ -5,6 +5,7 @@ import 'package:flutter_chat_app/constants/collections.dart';
 import 'package:flutter_chat_app/dialogs/popup.dart';
 import 'package:flutter_chat_app/models/pp_message.dart';
 import 'package:flutter_chat_app/models/user/pp_user_service.dart';
+import 'package:flutter_chat_app/services/contacts_event.dart';
 import 'package:flutter_chat_app/services/contacts_service.dart';
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/adapters.dart';
@@ -25,37 +26,63 @@ class ConversationService {
       .collection(Collections.User).doc(receiver)
       .collection(Collections.Messages);
 
-  StreamSubscription? listener;
+  StreamSubscription? _messagesListener;
+  StreamSubscription? _contactsEventListener;
+
   // < nickname, hiveBox>
   Map<String, Box<PpMessage>> _conversationsBoxes = {};
 
+
   login() async {
-    //open hive conversation boxes
+    // open hive conversation boxes
+    _userService.authValidate(where: 'conversation service');
     for (var contactNickname in _contactsService.currentContactNicknames) {
-      final box = await Hive.openBox<PpMessage>(_getHiveConversationKey(contactNickname));
-      _conversationsBoxes[contactNickname] = box;
+      await _addConversation(contactNickname);
     }
 
+    // listen for contacts list modification
+    _contactsEventListener = _contactsService.contactsEventStream.listen((event) {
+      switch (event.type) {
+        case ContactsEventTypes.add:
+          _addConversation(event.contactNickname);
+          break;
+        case ContactsEventTypes.remove:
+          _removeConversation(event.contactNickname);
+          break;
+      }
+    }, onError: (error) {
+      print('in _contactsEventListener error:');
+      print(error);
+    });
+
     // listen for coming messages - add to hive and delete from firestore
-    listener = messagesCollectionRef.snapshots().listen((event) async {
+    _messagesListener = messagesCollectionRef.snapshots().listen((event) async {
       for (var change in event.docChanges) {
-        try {
+        if (change.type == DocumentChangeType.added) {
           await _addMessageToHive(PpMessage.fromDB(change.doc));
           await change.doc.reference.delete();
-        } catch (error) {
-          _popup.sww(text: 'receiving message error');
         }
       }
+    }, onError: (error) {
+      print('in _messagesListener error:');
+      print(error);
     });
+    print('conversation service initialized');
   }
 
+  //TODO: add first message to conversation!
+
+
   logout() async {
-    for (var key in _conversationsBoxes.keys) {
-      _conversationsBoxes.remove(key);
+    await _contactsEventListener!.cancel();
+    for (var box in _conversationsBoxes.values) {
+      await box.close();
+      print('box closed');
     }
     await Hive.close();
     _conversationsBoxes = {};
-    listener!.cancel();
+    _messagesListener!.cancel();
+    print('conversation service loogged out');
   }
 
   Box<PpMessage> getConversationBox(String contactNickname) {
@@ -70,6 +97,17 @@ class ConversationService {
     } catch (error) {
       _popup.sww();
     }
+  }
+
+  _addConversation(String contactNickname) async {
+    final box = await Hive.openBox<PpMessage>(_getHiveConversationKey(contactNickname));
+    _conversationsBoxes[contactNickname] = box;
+    print('ready add conversation: $contactNickname');
+  }
+
+  _removeConversation(String contactNickname) async {
+    await Hive.box(_getHiveConversationKey(contactNickname)).deleteFromDisk();
+    _conversationsBoxes.remove(contactNickname);
   }
 
   _addMessageToHive(PpMessage message) async {
@@ -95,7 +133,8 @@ class ConversationService {
   }
 
   deleteBox(String receiverNickname) async {
-    //TODO: delete when delete account
+    //TODO: delete when delete conversation
+    //TODO: when delete conversation send also notifications about it and implement resolve it
     _conversationsBoxes.remove(receiverNickname);
     await Hive.box(_getHiveConversationKey(receiverNickname)).deleteFromDisk();
     print('deleted');
@@ -103,6 +142,7 @@ class ConversationService {
 
   deleteHiveData() async {
     //TODO: delete when delete account
+    //TODO: when delete account send accountDeleted notifications to contacts
     await Hive.deleteFromDisk();
   }
